Updated import paths to support the connectors relocation
With the connectors moved, much of the code that depended on them didn't
have proper import paths. This change fixes those issues.
