+++ internal/plugin/connectors/tcp/mysql/authentication_handshake.go
@@ -4,7 +4,7 @@ import (
 	"net"
 
 	"github.com/cyberark/secretless-broker/internal/handlers/ssl"
-	"github.com/cyberark/secretless-broker/internal/proxyservice/tcp/mysql/protocol"
+	"github.com/cyberark/secretless-broker/internal/plugin/connectors/tcp/mysql/protocol"
 )
 
 /*
