+++ internal/plugin/connectors/tcp/mysql/connector.go
@@ -3,7 +3,7 @@ package mysql
 import (
 	"net"
 
-	"github.com/cyberark/secretless-broker/internal/proxyservice/tcp/mysql/protocol"
+	"github.com/cyberark/secretless-broker/internal/plugin/connectors/tcp/mysql/protocol"
 	"github.com/cyberark/secretless-broker/pkg/secretless/log"
 	"github.com/cyberark/secretless-broker/pkg/secretless/plugin/connector"
 )
