+++ internal/plugin/connectors/tcp/pg/backend.go
@@ -5,7 +5,7 @@ import (
 	"net"
 
 	"github.com/cyberark/secretless-broker/internal/handlers/ssl"
-	"github.com/cyberark/secretless-broker/internal/proxyservice/tcp/pg/protocol"
+	"github.com/cyberark/secretless-broker/internal/plugin/connectors/tcp/pg/protocol"
 )
 
 // handleSSL conditionally upgrades the backend connection to SSL depending on the
