+++ internal/plugin/connectors/tcp/proxy_service_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	loggerMock "github.com/cyberark/secretless-broker/internal/log/mock"
-	"github.com/cyberark/secretless-broker/internal/proxyservice/mock"
+	"github.com/cyberark/secretless-broker/internal/plugin/connectors/mock"
 )
 
 const testString1 = "good heavens"
