+++ internal/proxyservice/proxy_service.go
@@ -9,11 +9,11 @@ import (
 
 	"github.com/cyberark/secretless-broker/internal"
 	"github.com/cyberark/secretless-broker/internal/plugin"
+	httpproxy "github.com/cyberark/secretless-broker/internal/plugin/connectors/http"
+	sshproxy "github.com/cyberark/secretless-broker/internal/plugin/connectors/ssh"
+	sshagentproxy "github.com/cyberark/secretless-broker/internal/plugin/connectors/sshagent"
+	tcpproxy "github.com/cyberark/secretless-broker/internal/plugin/connectors/tcp"
 	v1 "github.com/cyberark/secretless-broker/internal/plugin/v1"
-	httpproxy "github.com/cyberark/secretless-broker/internal/proxyservice/http"
-	sshproxy "github.com/cyberark/secretless-broker/internal/proxyservice/ssh"
-	sshagentproxy "github.com/cyberark/secretless-broker/internal/proxyservice/sshagent"
-	tcpproxy "github.com/cyberark/secretless-broker/internal/proxyservice/tcp"
 	v2 "github.com/cyberark/secretless-broker/pkg/secretless/config/v2"
 	logapi "github.com/cyberark/secretless-broker/pkg/secretless/log"
 	plugin2 "github.com/cyberark/secretless-broker/pkg/secretless/plugin"
